+++ Main/ClassAE.py
@@ -13,7 +13,7 @@ import matplotlib.pyplot as plt
 from tensorflow.keras.layers import Input, Conv2D, MaxPool2D, AveragePooling2D, UpSampling2D, Conv2DTranspose
 from tensorflow.keras.optimizers import Adam
 # Local codes
-from SampleFlows.ParentClass import Model
+from ParentClass import Model
 
 
 # Uncomment if keras does not run
