+++ Main/ClassPOD.py
@@ -1,6 +1,6 @@
 import numpy as np
 from numpy import linalg as LA
-from Model.ParentClass import Model
+from ParentClass import Model
 import matplotlib.pyplot as plt
 
 import os
